requirements.txt
appdirs==1.4.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
ECPy==1.2.5
et-xmlfile==1.1.0
future==0.18.3
httpq==1.2.0
idna==3.4
Levenshtein==0.21.0
libnum==1.7.1
mitm==1.5.0
openpyxl==3.1.2
pbr==5.9.0
protobuf==4.23.0
pwnedpasswords==2.0.0
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.1.1
rapidfuzz==3.0.0
requests==2.30.0
toolbox==1.11.0
tqdm==4.65.0
urllib3==2.0.2
